Multitapering

Multitapering is a popular technique for estimating the spectral density function of a signal in time. Multitapering can also be extended to spatial processes, including both random fields [3], point processes [4] and multivariate processes which are a mixture of the two [5].

Data format

Data should be passed into multitapering functions as a either PointSet or GeoTable objects. If the data is multivariate, is should be a Tuple of such objects. If the data is a GeoTable, and the domain of this data is a PointSet, then this is interpreted as a marked point process, where the first column of the references information is the mark. If the domain is a CartesianGrid, then this is interpreted as a random field sampled on that grid. Again we take only the first column, so you should pass multivariate processes as separate GeoTable objects, even if they are recorded at the same location.

Multivariate processes

SpatialMultitaper.multitaper_estimate — Function

multitaper_estimate(data, region; nfreq, fmax, tapers, mean_method = DefaultMean())

Computes the multitaper spectral estimate from a tapered dft.

Arguments:

data: The data to estimate the spectrum from.

region: The region to estimate the spectrum from.

nfreq::NTuple{D,Int}: The number of frequencies in each dimension.

fmax::NTuple{D,Real}: The maximum frequency in each dimension.

tapers: A tuple of taper functions.

mean_method: The method to estimate the mean. Default is DefaultMean, but can use KnownMean(x) to specify this if known.

Output:

The output is a SpectralEstimate object, with a freq and power field. Say nfreq = (n_1,...,n_D), then freq is a D dimensional NTuple of arrays of frequencies in each dimension. Power can be in the following forms:

If data is a single process, then power is a n_1 x ... x n_D array.

If data has P processes and is passed as an NTuple, then power is an n_1 x ... x n_D array of SMatrix{P,P,T,P*P}.

If data is a vector of P processes, then power is a P x P x n_1 x ... x n_D array.

In any case, indexing into a SpectralEstimate will provide a SpectralEstimate object with the same freq and a subset of the power array corresponding to the processes in the chosen index.

Mathematical background

References

[3]

A. Hanssen. Multidimensional multitaper spectral estimation. Signal Processing 58, 327–332 (1997).

[4]

T. A. Rajala, S. C. Olhede, J. P. Grainger and D. J. Murrell. What is the Fourier Transform of a Spatial Point Process? IEEE Transactions on Information Theory 69, 5219–5252 (2023).

[5]

J. P. Grainger, T. A. Rajala, D. J. Murrell and S. C. Olhede. Spectral estimation for spatial point processes and random fields, arXiv preprint arXiv:2312.10176 (2023).