Reference

Contents

Reference

Contents

Index

Index

SpatialMultitaper.SlepianSolver.LinearOperator

SpatialMultitaper.L2_inner_product_interpolated

SpatialMultitaper.SlepianSolver.compute_eigenfunctions

SpatialMultitaper.SlepianSolver.make_concentration_operator

SpatialMultitaper.box2sides

SpatialMultitaper.centerpad

SpatialMultitaper.choose_freq_1d

SpatialMultitaper.choose_freq_res

SpatialMultitaper.dft2spectralmatrix

SpatialMultitaper.dft2spectralmatrix

SpatialMultitaper.downsample

SpatialMultitaper.fft_anydomain

SpatialMultitaper.freq_downsample_index

SpatialMultitaper.freq_downsample_startindex

SpatialMultitaper.grid2side

SpatialMultitaper.interpolated_taper_family

SpatialMultitaper.make_tapers

SpatialMultitaper.multitaper_estimate

SpatialMultitaper.newweight

SpatialMultitaper.nufft1d1_anydomain

SpatialMultitaper.nufft1d1_anydomain!

SpatialMultitaper.nufft1d1_anydomain_precomp

SpatialMultitaper.nufft2d1_anydomain

SpatialMultitaper.nufft2d1_anydomain!

SpatialMultitaper.nufft2d1_anydomain_precomp

SpatialMultitaper.nufft3d1_anydomain

SpatialMultitaper.nufft3d1_anydomain!

SpatialMultitaper.nufft3d1_anydomain_precomp

SpatialMultitaper.nufft_anydomain

SpatialMultitaper.optimaltapers

SpatialMultitaper.pad

SpatialMultitaper.partial_K

SpatialMultitaper.partial_spectra

SpatialMultitaper.points2coords

SpatialMultitaper.reprocess

SpatialMultitaper.rescale_points

SpatialMultitaper.sdf2K

SpatialMultitaper.side2grid

SpatialMultitaper.sin_taper_family

SpatialMultitaper.tapered_dft

SpatialMultitaper.tapered_dft

SpatialMultitaper.unwrap_fft_output

SpatialMultitaper.unwrap_index

SpatialMultitaper.upsample

SpatialMultitaper.L2_inner_product_interpolated — Method

interp_ip_2(h::AbstractArray{T,D}, g::AbstractArray{T,D}, grid::CartesianGrid) where {T,D}

The L₂ inner product of interpolated sequences g and h recorded on grid grid.

SpatialMultitaper.box2sides — Method

box2sides(box::Box)

Converts a box to a tuple of tuples, whose j'th entry is thej`th coordinate of the minimum and maximum of the box.

SpatialMultitaper.centerpad — Method

centerpad(x, i)

Pads x with zeros on all sides by i elements. For example, if x = 1:2 and i = 1, then the output is [0, 1, 2, 0].

SpatialMultitaper.choose_freq_1d — Method

choose_freq_1d(nfreq, fmax)

Returns the frequencies for our fft interfaces, with nfreq frequencies, and fmax as the fmax frequency.

SpatialMultitaper.choose_freq_res — Method

choose_freq_res(n, nfreq; maxitr=100)

Find the smallest integer oversample such that nfreq*oversample ≥ n.

SpatialMultitaper.dft2spectralmatrix — Method

dft2spectralmatrix(J_n::Array)

Computes the spectral matrix from the DFTs, assuming that the DFTs are stored as one large array which is P x M x n1 x ... x nD.

SpatialMultitaper.dft2spectralmatrix — Method

dft2spectralmatrix(J_n::NTuple{P, Array{T, N}}) where {P, T, N}

Computes the spectral matrix from the DFTs, assuming that the DFTs are stored as a tuple of P arrays of size n1 x ... x nD x M.

SpatialMultitaper.downsample — Method

downsample(x::AbstractArray{T,D}, spacing) where {D,T}

Down sample an AbstractArray. Specify spacing as an Int, for same spacing in all dims. Specify spacing as an NTuple{D,Int} for different spacing. Specify spacing=nothing to just return x.

SpatialMultitaper.fft_anydomain — Method

fft_anydomain(x::Array, grid::Grid, nfreq, fmax; kwargs...)

Apply an fft to data stored in x given that x is recorded on a grid grid. The output is at frequencies determined by nfreq and fmax. x can have more dimensions than grid but the first dimensions must be the same as the grid (up to the size of the grid).

SpatialMultitaper.freq_downsample_index — Method

freq_downsample_index(nfreq, oversample)

Returns the indices required to downsample the oversampled frequencies with a given oversampling. This assumes that the frequencies are fftshifted.

Arguments

nfreq::Int: The number of frequencies for the desired output.

oversample::Int: The oversampling factor used.

SpatialMultitaper.freq_downsample_startindex — Method

freq_downsample_startindex(nfreq,oversample)

Returns the starting index for the oversampled frequencies to downsample.

SpatialMultitaper.grid2side — Method

grid2side(g::CartesianGrid)

Converts a grid to a tuple of the location of the centers for each side.

SpatialMultitaper.interpolated_taper_family — Method

interpolated_taper_family(raw_tapers, grid; freq_res=size(grid))

Constructs a multitaper object from multiple taper discrete impulse responses on a grid.

Arguments

raw_tapers: A Vector of AbstractArrays containing the taper impulse response.

grid: The grid corresponding to the tapers.

freq_res: Optional named argument specifying the upsampling for computing the transfer function.

Frequency sampling

The frequency grid on which the transfer function will be sampled has:

- length `freq_res`.
- interval `unitless_spacing(grid)/freq_res`.

SpatialMultitaper.make_tapers — Method

make_tapers(region; bandwidth, threshold = 0.99, space_res = 100, freq_res = 500, freq_generate_res = 500, ntapers_max = 30)

A function to create tapers for a given region.

Arguments

region: The region to create the tapers on.

bandwidth: The bandwidth of the tapers

threshold: The threshold for the number of tapers to use.

space_res: The resolution of the grid to create the tapers on.

freq_res: The resolution of the frequency grid to compute the transfer function on.

freq_generate_res: The resolution of the frequency grid to generate the tapers on.

ntapers_max: The maximum number of tapers to generate.

As a rough rule of thumb, setting the bandwidth to an integer multiple (say x) of a side length of the bounding square of a region will produce x^d tapers (note this is a very crude approximation, but you will find out how many good tapers you can make when you create them).

SpatialMultitaper.multitaper_estimate — Method

multitaper_estimate(data, region; nfreq, fmax, tapers, mean_method = DefaultMean())

Computes the multitaper spectral estimate from a tapered dft.

Arguments:

data: The data to estimate the spectrum from.

region: The region to estimate the spectrum from.

nfreq::NTuple{D,Int}: The number of frequencies in each dimension.

fmax::NTuple{D,Real}: The maximum frequency in each dimension.

tapers: A tuple of taper functions.

mean_method: The method to estimate the mean. Default is DefaultMean, but can use KnownMean(x) to specify this if known.

Output:

The output is a SpectralEstimate object, with a freq and power field. Say nfreq = (n_1,...,n_D), then freq is a D dimensional NTuple of arrays of frequencies in each dimension. Power can be in the following forms:

If data is a single process, then power is a n_1 x ... x n_D array.

If data has P processes and is passed as an NTuple, then power is an n_1 x ... x n_D array of SMatrix{P,P,T,P*P}.

If data is a vector of P processes, then power is a P x P x n_1 x ... x n_D array.

In any case, indexing into a SpectralEstimate will provide a SpectralEstimate object with the same freq and a subset of the power array corresponding to the processes in the chosen index.

SpatialMultitaper.newweight — Method

newweight(h, freq_downsampling)

Necessary because eigs gives normalised vector under the downsampling.

SpatialMultitaper.nufft1d1_anydomain! — Method

nufft1d1_anydomain!(output_storage, input_data, nfreq, fmax, iflag, eps; kwargs...)

In place version of nufft1d1_anydomain. Use nufft1d1_anydomain_precomp to precompute the input data and memory required.

SpatialMultitaper.nufft1d1_anydomain — Method

nufft1d1_anydomain(interval, nfreq, fmax, xj, cj, iflag, eps; kwargs...)

Computes the approximate dft of a 1d function using the nufft algorithm, with points in any interval, at frequencies defined by fmax and nfreq. xj are the points coordinates and cj are the function values. Set iflag<0 for -i in exponent.

SpatialMultitaper.nufft1d1_anydomain_precomp — Method

nufft1d1_anydomain_precomp(interval, nfreq, fmax, xj, cj)

Precomputes the input data and memory required for nufft1d1_anydomain!.

SpatialMultitaper.nufft2d1_anydomain! — Method

nufft2d1_anydomain!(output_storage, input_data, nfreq, fmax, iflag, eps; kwargs...)

In place version of nufft2d1_anydomain. Use nufft2d1_anydomain_precomp to precompute the input data and memory required.

SpatialMultitaper.nufft2d1_anydomain — Method

nufft2d1_anydomain(box, nfreq, fmax, xj, yj, cj, iflag, eps; kwargs...)

Computes the approximate dft of a 2d function using the nufft algorithm, with points in any box, at frequencies defined by fmax and nfreq. box here should be a tuple of intervals, one for each dimension. Similarly for fmax and nfreq. xj, yj are the points coordinates and cj are the function values. Set iflag<0 for -i in exponent.

SpatialMultitaper.nufft2d1_anydomain_precomp — Method

nufft2d1_anydomain_precomp(box, nfreq, fmax, xj, yj, cj)

Precomputes the input data and memory required for nufft2d1_anydomain!.

SpatialMultitaper.nufft3d1_anydomain! — Method

nufft3d1_anydomain!(output_storage, input_data, nfreq, fmax, iflag, eps; kwargs...)

In place version of nufft3d1_anydomain. Use nufft3d1_anydomain_precomp to precompute the input data and memory required.

SpatialMultitaper.nufft3d1_anydomain — Method

nufft3d1_anydomain(cube, nfreq, fmax, xj, yj, zj, cj, iflag, eps; kwargs...)

Computes the approximate dft of a 3d function using the nufft algorithm, with points in any cube, at frequencies defined by fmax and nfreq. cube here should be a tuple of intervals, one for each dimension. Similarly for fmax and nfreq. xj, yj, zj are the points coordinates and cj are the function values. Set iflag<0 for -i in exponent.

SpatialMultitaper.nufft3d1_anydomain_precomp — Method

nufft3d1_anydomain_precomp(cube, nfreq, fmax, xj, yj, zj, cj)

Precomputes the input data and memory required for nufft3d1_anydomain!.

SpatialMultitaper.nufft_anydomain — Method

nufft_anydomain(region, nfreq, fmax, points::PointSet, c, iflag, eps; kwargs...)

Computes the approximate dft of a function using the nufft algorithm, with points in any region, at frequencies defined by fmax and nfreq. Currently only 1d, 2d and 3d are supported.

SpatialMultitaper.optimaltapers — Method

optimaltapers(region::Meshes.GeometryOrDomain, grid::CartesianGrid; freq_region::Meshes.GeometryOrDomain, ntapers::Int, freq_res, freq_downsample=nothing, real_tapers=true, tol=0.0)

Function to compute the optimal tapers for a given region, grid and region in frequency.

Arguments:

region: The spatial observation region. Should be of type Geometry or Domain.

grid: A CartesianGrid containing the region on which we wish to have a taper.

freq_region: The region in frequency we want to concentrate the taper on. Typically a Ball centered at zero.

ntapers: The number of desired tapers.

freq_res: The oversampling to be used in frequency.

real_tapers: Optional argument to decided if real or complex tapers should be provided. Default is true which provides real tapers.

tol: Optional argument passed to the eigs function of Arpack. You likely need to play with this.

Note about errors:

If you have an error, it is likely a convergence problem. Try setting a larger value for tol.

SpatialMultitaper.pad — Method

pad(x::AbstractArray{T,D}, n::NTuple{D,Int}) where {T,D}
pad(x::AbstractArray{T,D}, n::Int) where {T,D}

Pad x with zeros to size n. Note that size(x) ≤ n must hold, and the new array is of size n not size(x).+n. If an integer is provided for n, it is interpreted as ntuple(d->n, Val{D}()).

SpatialMultitaper.partial_K — Method

partial_K(data, region; radii, nfreq, fmax, tapers, indices, upscalefactor)

Computes the partial K function from the data at radii radii. Default is to compute this for all pairs of indices conditional on any index not included. Alternatively, pass a vector of indices. If this is a vector of Tuple{Int,Int}, then this is computed partial on every other index. If this is a Tuple{Int,Int,AbstractVector{Int},AbstractVector{Int}}, then this is computed partial on the specified indices, i.e. The residual of index[1] partial index[3] with index[2] partial index[4]. Setting upscalefactor to nothing does no upscaling. Otherwise, set to an Int or an NTuple{D,Int} to upscale the spectra for the computation of the partial K.

SpatialMultitaper.partial_spectra — Method

partial_spectra(x::Matrix)

If only a matrix x is passed, computes the partial spectra for all indices. The diagonal elements are the spectra of residuals given all other processes The i j th element (for i ≠ j) is the partial coherence between the ith and jth process given all other processes not including i and j.

partial_spectra(x::Matrix, i1::Int, i2::Int, c1, c2)

If specific indices are requested, computes the partial spectra for the i1th index conditioned on the indices in c1 vs the i2th index conditioned on the indices in c2.

SpatialMultitaper.points2coords — Method

points2coords(points::PointSet)

Converts PointSet to a tuple of view of reinterpreted arrays, whose j'th entry is thej`th coordinate of each of the points.

SpatialMultitaper.reprocess — Method

reprocess(h_large, grid, freq_downsample)

Undoes the downsampling and padding.

SpatialMultitaper.rescale_points — Method

rescale_points(x, nfreq, fmax, interval; max_oversample = 100)

Rescales points to be in the interval [-π, π] and returns the oversampling factor. Oversampling factor is the smallest integer such that the interval is rescaled to fit in [-π, π], and the corresponding fmax frequency is a multiple of fmax. Interval just needs to have minimum and maximum defined.

SpatialMultitaper.sdf2K — Method

sdf2K(freq, spectra::AbstractArray{T,D}, r) where {D,T}

Computes the K function from the spectral density function.

SpatialMultitaper.side2grid — Method

side2grid(sides)

inverse of grid2side

SpatialMultitaper.sin_taper_family — Method

make_sin_tapers(lbw, region::Box, gridsize)

Construct the sin tapers for a given region and gridsize.

Arguments:

ntapers: The number of tapers.

region: The region to construct the tapers on. Must be a box.

Background:

The sin tapers on a unit interval are defined as:

\[h_m(x) = \sqrt{2} * \sin(πxm)\]

SpatialMultitaper.tapered_dft — Method

tapered_dft(data::AbstractVector{<:Union{GeoTable,PointSet}}, tapers, nfreq, fmax, region, mean_method::AbstractVector{<:MeanEstimationMethod})

Compute the tapered discrete Fourier transform of a collection of data sets. If we have n_1,...,n_D frequencies, and M tapers and P processes, this returns an array of size P x M x n_1 x ... x n_D. Note that for small numbers of processes, it is more efficient to use the NTuple method.

SpatialMultitaper.tapered_dft — Method

tapered_dft(data::NTuple{P,Union{GeoTable, PointSet}}, tapers, nfreq, fmax, region, mean_method::NTuple{P,MeanEstimationMethod}) where {P}

Compute the tapered discrete Fourier transform of a collection of data sets. If we have n1,...,nD frequencies, and M tapers and P processes, this returns a Tuple of size P, whose entries are arrays of size n_1 x ... x n_D x M.

SpatialMultitaper.unwrap_fft_output — Method

unwrap_fft_output(J::AbstractArray, fmaxrel::NTuple{D, Int}) where {D}

Copies the result of the fft to give something which goes up to some multiple of the nyquist frequency (whilst keeping the number of output frequencies).

SpatialMultitaper.unwrap_index — Method

unwrap_index(n, a)

The indices required to unwrap the fft output to go up to n with a given oversampling a.

SpatialMultitaper.upsample — Method

upsample(x::AbstractArray, grid::CartesianGrid, freq_downsample)

Upsamples an array using linear interpolation, checking the grid size against x. Note this is specific to downsample as used here, and shouldn't be used for general upsampling.

SpatialMultitaper.SlepianSolver.LinearOperator — Type

LinearOperator

A linear operator. Assumes that inputspace and outputspace are defined. One should also define mul!(y, L::LinearOperator, x) and preallocate_output(L::LinearOperator, x) for efficiency.

SpatialMultitaper.SlepianSolver.compute_eigenfunctions — Method

compute_eigenfunctions(R::AbstractArray,K::AbstractArray,ntapers=1;imag_tol=1e-5,real_tapers=true, tol=0.0)

Computes the eigenfunctions of the concentration operator. Provide real_tapers=true for real valued tapers.

SpatialMultitaper.SlepianSolver.make_concentration_operator — Function

make_concentration_operator(R::AbstractArray,K::AbstractArray,real_tapers=Val{true}())

Constructs the concentration operator. Provide real_tapers=Val{true}() for real valued tapers. Choose real_tapers=Val{false}() for complex valued tapers.